+++ src/test/java/com/jardoapps/changelog/merge/driver/ChangelogMergerTest.java
@@ -167,6 +167,56 @@ class ChangelogMergerTest {
 	/**
 	 * Model situation:
 	 *   <ul>
+	 *     <li>Maintaining 2 versions: 1.0.x and 2.0.x</li>
+	 *     <li>A fix "Fix 1" has been made in version 1.0.1-SNAPSHOT and merged to 2.0.1-SNAPSHOT</li>
+	 *     <li>Version 2.0.1 has been released (but 1.0.1 still has not)</li>
+	 *     <li>Fix "Fix 1" should not be repeated again in the unreleased section of ours.
+	 *         The "Fixed" section in ours would be blank, and therefore should not be added.
+	 *     </li>
+	 *   </ul>
+	 */
+	@Test
+	void testMerge_dontAddBlankSections() {
+
+		Changelog ourChangelog = Changelog.builder()
+				.name("Changelog")
+				.unreleasedVersion(Changelog.Version.builder()
+						.name("2.0.2")
+						.releaseDate("[SNAPSHOT]")
+						.build())
+				.releasedVersion(Changelog.Version.builder()
+						.name("2.0.1")
+						.releaseDate("2024-04-12")
+						.section(Changelog.Section.builder()
+								.name("Fixed")
+								.line("")
+								.line("- Fix 1")
+								.build())
+						.build())
+				.build();
+
+		Changelog theirChangelog = Changelog.builder()
+				.name("Changelog")
+				.unreleasedVersion(Changelog.Version.builder()
+						.name("1.0.1")
+						.releaseDate("[SNAPSHOT]")
+						.section(Changelog.Section.builder()
+								.name("Fixed")
+								.line("")
+								.line("- Fix 1")
+								.build())
+						.build())
+				.build();
+
+		Changelog mergedChangelog = changelogMerger.merge(ourChangelog, theirChangelog);
+
+		Version unreleasedVersion = mergedChangelog.getUnreleasedVersion();
+		assertThat(unreleasedVersion.getSections()).hasSize(0);
+	}
+
+	/**
+	 * Model situation:
+	 *   <ul>
 	 *     <li>Maintaining 3 versions: 1.0.0, 2.0.0, 3.0.0</li>
 	 *     <li>A fix has been made in version 1.0.1-SNAPSHOT and merged to all other versions.</li>
 	 *     <li>Version 1.0.1 has been released and merged to 2.0.1-SNAPSHOT. The fix in 2.0.1-SNAPSHOT has been marked with [from `1.0.1`].</li>